restaurant-microservices/shared/requirements-base.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.0
PyJWT>=2.8.0
requests>=2.31.0
redis>=5.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
